Corrected version, with romanian input data
We missed the part where it stated that the input had to be romanian, so I had to create a new input file and re-run the code. The password is still the same, however. I sincerely apologize for the inconvenience, in case you had already started working on deciphering the data.